+++ messages/en.json
@@ -6498,6 +6498,17 @@
     "theater_balance_ok": "Every year balances",
     "theater_balance_warn": "Imbalance found, see the warnings afterwards",
     "theater_writing": "Writing to the journal ...",
+    "reveal_eyebrow": "Your history",
+    "reveal_title": "{years, plural, one {Your history is in place.} other {# years of history, in place.}}",
+    "reveal_stats": "{vouchers} vouchers · {accounts} accounts · {counterparties} counterparties",
+    "reveal_tie_ok": "Every year balances: 0.00 kr off",
+    "reveal_skipped": "{count, plural, one {# voucher was skipped, see details below.} other {# vouchers were skipped, see details below.}}",
+    "reveal_meta_period_created": "Fiscal year created",
+    "reveal_meta_period_existing": "Existing fiscal year",
+    "reveal_meta_ib": "opening balances imported",
+    "reveal_bridge": "The history is in place. What's missing is the present: the bank.",
+    "reveal_cta_bank": "Connect the bank",
+    "reveal_cta_open": "Open Accounted",
     "export_title": "Export",
     "export_subtitle": "Download your bookkeeping as a SIE file or back it up to Google Drive",
     "tab_import": "Import",
