+++ messages/sv.json
@@ -6498,6 +6498,17 @@
     "theater_balance_ok": "Varje år balanserar",
     "theater_balance_warn": "Obalans hittad, se varningarna efteråt",
     "theater_writing": "Skriver till journalen ...",
+    "reveal_eyebrow": "Din historia",
+    "reveal_title": "{years, plural, one {Historiken är på plats.} other {# år av historia, på plats.}}",
+    "reveal_stats": "{vouchers} verifikat · {accounts} konton · {counterparties} motparter",
+    "reveal_tie_ok": "Varje år balanserar: 0,00 kr i diff",
+    "reveal_skipped": "{count, plural, one {# verifikat hoppades över, se detaljer nedan.} other {# verifikat hoppades över, se detaljer nedan.}}",
+    "reveal_meta_period_created": "Räkenskapsår skapat",
+    "reveal_meta_period_existing": "Befintligt räkenskapsår",
+    "reveal_meta_ib": "ingående balanser importerade",
+    "reveal_bridge": "Historiken är på plats. Det som saknas är nuet: banken.",
+    "reveal_cta_bank": "Koppla banken",
+    "reveal_cta_open": "Öppna Accounted",
     "export_title": "Exportera",
     "export_subtitle": "Ladda ner bokföringen som SIE-fil eller säkerhetskopia till Google Drive",
     "tab_import": "Importera",
